sns게시물 javascript, css 정리 및 수정

+++ harulook/src/main/webapp/resources/js/boardDetail.js
@@ -2,20 +2,23 @@
 function boardDetail(msg) {
 	
 	console.log('boardDatil : ',msg);
+	
+	/* 게시물 상세보기 이미지 */
 	var snsDetailImg = '';
 	snsDetailImg += '<img alt="no image" style="width: 580px;" onError="this.src=\'resources/files/images/defaut.jpg\';" src="'+msg.board.snsBoardImg+'">';
-	
 	$('#snsDetailImg').html(snsDetailImg);
+	
+	/* 게시물 상세정보 */
 	var snsDetailContent = '';
 	snsDetailContent += '<div style="float: left;">';
 	snsDetailContent += '<h4>'+msg.board.userNick+'(<a href="./boardTagSearch?snsBoardAge=&snsBoardLoc=&snsBoardSize=&snsBoardTall=&snsBoardWeather=&userId='+msg.board.userId+'">'+msg.board.userId+'</a>)</h4>';
 	snsDetailContent += '</div>';
-	snsDetailContent += '<button id="followInsertButton">팔로우</button>';
+	snsDetailContent += '<button id="followInsertButton" class="btn btn-default">팔로우</button>';
 	snsDetailContent += '<span id="followCheckOverlap">이미등록된친구</span>';
 	snsDetailContent += '<span id="Blank">&nbsp;</span>';
 	if(msg.board.userId == $('#sessionUserId').val()) {
 		snsDetailContent += '<div class="menu">';
-		snsDetailContent += '<div id="item"><img src="resources/files/images/more_icon.png" style="width: 30px;"></div>';
+		snsDetailContent += '<div id="item">&nbsp<i class="fa fa-ellipsis-h" style="font-size:24px"></i></div>';
 		snsDetailContent += '<div id="submenu">';
 		snsDetailContent += '<a href="./boardUpdate?boardNo='+msg.board.snsBoardNo+'">수정</a>';
 		snsDetailContent += '<a href="./boardDelete?boardNo='+msg.board.snsBoardNo+'">삭제</a>';
@@ -24,7 +27,6 @@ function boardDetail(msg) {
 	}
 	snsDetailContent += '<hr>';
 	snsDetailContent += '<div>';
-	
 	
 	if(msg.board.snsBoardWeather != '') {
 		snsDetailContent += '<span>날씨&nbsp:&nbsp'+msg.board.snsBoardWeather+'</span><br>';
@@ -59,7 +61,7 @@ function boardDetail(msg) {
 		case '빈티지' :
 			snsStyleValue = 'style_03';
 			break;
-		case '스트리트' :
+		case '스트릿' :
 			snsStyleValue = 'style_04';
 			break;
 		case '댄디' :
@@ -84,7 +86,7 @@ function boardDetail(msg) {
 			snsStyleValue = '';
 		break;
 		}
-		snsDetailContent += '<a href="./boardTagSearch?snsBoardAge=&snsBoardLoc=&snsBoardSize=&snsBoardTall=&snsBoardWeather=&styleValue='+snsStyleValue+'">#'+msg.snsStyle[i]+'</a>&nbsp';
+		snsDetailContent += '<a href="./boardTagSearch?snsBoardAge=&snsBoardLoc=&snsBoardSize=&snsBoardTall=&snsBoardWeather=&userId=&styleValue='+snsStyleValue+'">#'+msg.snsStyle[i]+'</a>&nbsp';
 	}
 	for(var i=0; i<msg.snsColor.length; i++) {
 		var snsColorValue = '';
@@ -132,7 +134,7 @@ function boardDetail(msg) {
 			snsColorValue = '';
 		break;
 		}
-		snsDetailContent += '<a href="./boardTagSearch?snsBoardAge=&snsBoardLoc=&snsBoardSize=&snsBoardTall=&snsBoardWeather=&colorValue='+snsColorValue+'">#'+msg.snsColor[i]+'</a>&nbsp';
+		snsDetailContent += '<a href="./boardTagSearch?snsBoardAge=&snsBoardLoc=&snsBoardSize=&snsBoardTall=&snsBoardWeather=&userId=&colorValue='+snsColorValue+'">#'+msg.snsColor[i]+'</a>&nbsp';
 	}
 	for(var i=0; i<msg.snsSituation.length; i++) {
 		var snsSituationValue = '';
@@ -162,35 +164,41 @@ function boardDetail(msg) {
 			snsSituationValue = '';
 		break;
 		}
-		snsDetailContent += '<a href="./boardTagSearch?snsBoardAge=&snsBoardLoc=&snsBoardSize=&snsBoardTall=&snsBoardWeather=&situationValue='+snsSituationValue+'">#'+msg.snsSituation[i]+'</a>&nbsp';
+		snsDetailContent += '<a href="./boardTagSearch?snsBoardAge=&snsBoardLoc=&snsBoardSize=&snsBoardTall=&snsBoardWeather=&userId=&situationValue='+snsSituationValue+'">#'+msg.snsSituation[i]+'</a>&nbsp';
 	}
 	snsDetailContent += '</div>';
 	
 	$('#snsDetailContent').html(snsDetailContent);
 	
-	var snsDetailComment = '';
+	/* 게시물 추천 유무와 추천수 */
+	var snsDetailLike = '';
 	
-	snsDetailComment += '<div>';
-	snsDetailComment += '<div id="like">';
+	snsDetailLike += '<div>';
+	snsDetailLike += '<div id="like">';
 	if($('#sessionUserLevel').val() == '일반회원') {
 		
-		snsDetailComment += '<button id="likeBtn" class="btn btn-link">';
+		snsDetailLike += '<button id="likeBtn" class="btn btn-link">';
 		console.log('likeToggle : ',msg.likeToggle);
+		// 이전에 해당 게시물을 추천한 경우, 아이콘을 분홍색으로 표시
 		if(msg.likeToggle == true) {
-			snsDetailComment += '<i id="likeToggle" class="material-icons center" style="color: #FFB2F5; font-size:24px;">thumb_up</i>';
-		} else {
-			snsDetailComment += '<i id="likeToggle" class="material-icons center" style="color: rgba(0,0,0,0.3); font-size:24px;">thumb_up</i>';
-		}
-		snsDetailComment += '</button>';
-		
+			snsDetailLike += '<i id="likeToggle" class="material-icons center" style="color: #FFB2F5; font-size:24px;">thumb_up</i>';
 
+		// 이전에 해당 게시물을 추천한 적이 없는 경우, 아이콘을 회색으로 표시
+		} else {
+			snsDetailLike += '<i id="likeToggle" class="material-icons center" style="color: rgba(0,0,0,0.3); font-size:24px;">thumb_up</i>';
+		}
+		snsDetailLike += '</button>';
+		
 	} else {
-		snsDetailComment += '<i class="material-icons center" style="color: rgba(0,0,0,0.3); font-size:24px;">thumb_up</i>';
+		snsDetailLike += '<i class="material-icons center" style="color: rgba(0,0,0,0.3); font-size:24px;">thumb_up</i>';
 	}
 	
-	snsDetailComment += '<span class="center" id="snsLikeCount">'+msg.board.snsLikeCount+'개&nbsp;&nbsp;</span>';
-	snsDetailComment += '</div>';
+	snsDetailLike += '<span class="center" id="snsLikeCount">'+msg.board.snsLikeCount+'개&nbsp;&nbsp;</span>';
+	snsDetailLike += '</div>';
+	$('#snsDetailLike').html(snsDetailLike);
 	
+	/* 게시물 댓글 정보 */
+	var snsDetailComment = '';
 	for(var i=0; i<msg.commentList.length; i++) {
 		snsDetailComment += '<span>'+msg.commentList[i].userId+': '+msg.commentList[i].snsCommentContent+'</span>';
 		if(msg.commentList[i].userId == $('#sessionUserId').val()) {
@@ -201,52 +209,62 @@ function boardDetail(msg) {
 	snsDetailComment += '</div>';
 	$('#snsDetailComment').html(snsDetailComment);
 	
+	
 	var snsDetailCommentControll = '';
 	snsDetailCommentControll += '<input type="text" id="commentValue">';
 	snsDetailCommentControll += '<button type="button" id="commentBtn">등록</button>';
 	$('#snsDetailCommentControll').html(snsDetailCommentControll);
 	
+	/* 댓글 등록 버튼 클릭 */
 	$('#commentBtn').click(function(){
 		var snsCommentContent = $('#commentValue').val();
-		var commentInsertRequest = $.ajax({
-			url: './commentInsert',
-			method: 'post',
-			data:{'snsCommentContent' : snsCommentContent
-				, 'userId' : $('#sessionUserId').val()
-				, 'snsBoardNo' : msg.board.snsBoardNo
-				}
-		});
-		commentInsertRequest.done(function(commentInsertData){
-			$.ajax({
-				url: './boardDetail',
-				method: 'get',
-				data: { 'boardNo' : msg.board.snsBoardNo},
-				datatype: 'json',
-				success: function(data){
-					snsDetailComment = '';
-					
-					snsDetailComment += '<div>';
-					for(var i=0; i<data.commentList.length; i++) {
-						snsDetailComment += '<span>'+data.commentList[i].userId+': '+data.commentList[i].snsCommentContent+'</span>';
-						if(data.commentList[i].userId == $('#sessionUserId').val()) {
-							snsDetailComment += '<a class="btn-sm" href="./commentDelete?snsCommentNo='+data.commentList[i].snsCommentNo+'&snsBoardNo='+msg.board.snsBoardNo+'">삭제</a>';
-						}
-						snsDetailComment += '<br>';
+		
+		console.log('commentValue : ',commentValue);
+		if(snsCommentContent == '') {
+			alert('댓글 내용을 입력하세요');
+		} else {
+			var commentInsertRequest = $.ajax({
+				url: './commentInsert',
+				method: 'post',
+				data:{'snsCommentContent' : snsCommentContent
+					, 'userId' : $('#sessionUserId').val()
+					, 'snsBoardNo' : msg.board.snsBoardNo
 					}
-					snsDetailComment += '</div>';
-					$('#snsDetailComment').html(snsDetailComment);
-				}
 			});
-		})
-		$('#commentValue').val('');
+			commentInsertRequest.done(function(commentInsertData){
+				$.ajax({
+					url: './boardDetail',
+					method: 'get',
+					data: { 'boardNo' : msg.board.snsBoardNo},
+					datatype: 'json',
+					success: function(data){
+						snsDetailComment = '';
+						
+						snsDetailComment += '<div>';
+						for(var i=0; i<data.commentList.length; i++) {
+							snsDetailComment += '<span>'+data.commentList[i].userId+': '+data.commentList[i].snsCommentContent+'</span>';
+							if(data.commentList[i].userId == $('#sessionUserId').val()) {
+								snsDetailComment += '<a class="btn-sm" href="./commentDelete?snsCommentNo='+data.commentList[i].snsCommentNo+'&snsBoardNo='+msg.board.snsBoardNo+'">삭제</a>';
+							}
+							snsDetailComment += '<br>';
+						}
+						snsDetailComment += '</div>';
+						$('#snsDetailComment').html(snsDetailComment);
+					}
+				});
+			})
+			$('#commentValue').val('');
+		}
 	});
 	
+	/* 게시물 수정,삭제 dropdown */
 	$('#submenu').hide();
 	
 	$('#item').click(function(){
 		$('#submenu').slideToggle(500);
 	});
 	
+	/* 예뻐요(추천) 버튼 클릭 */
 	$("#likeBtn").click(function(){
 		
 		console.log($('#sessionUserId').val());
